requirements.txt
discord.py
python-dotenv
logging
instaloader